+++ mmcls/models/utils/augment/identity.py
@@ -1,6 +1,6 @@
 import torch.nn.functional as F
 
-from .base_augment import BaseAugment
+from .augment import BaseAugment
 from .builder import AUGMENT
